pyproject.toml
[project]
name = "adaptive-finance-governance-agent"
version = "0.1.0"
description = "Multi-agent AI system for automated finance compliance with adaptive learning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langfuse>=3.9.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "a2a-sdk>=0.1.0",
    "mcp>=1.0.0",
    "sentence-transformers>=3.0.0",
    "openai>=1.50.0",
    "pillow>=10.0.0",
    "pdf2image>=1.17.0",
    "python-multipart>=0.0.9",
    "databricks-sql-connector>=4.2.1",
    "azure-storage-blob>=12.19.0",
]

[project.optional-dependencies]
# Document processing (OCR and text extraction)
ml = [
    "unstructured[pdf]>=0.10.0",
    "pytesseract>=0.3.10",
]

# Web UI
ui = [
    "streamlit>=1.39.0",
    "plotly>=5.24.0",
    "streamlit-agraph>=0.0.45",
]

# Development tools
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
]

# All optional dependencies
all = [
    "unstructured[pdf]>=0.10.0",
    "pytesseract>=0.3.10",
    "streamlit>=1.39.0",
    "plotly>=5.24.0",
    "streamlit-agraph>=0.0.45",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

